Fix: nvlists without NV_UNIQUE_NAME can cause kernel task crashes
Many of the ioctls in ZFS make use of the fnvlist_lookup_* family of functions,
which assert when the associated nvlist_lookup_* function returns non-zero. On
Linux, one consequence of this is that any program that issues an ioctl with a
misconfigured nvlist will hang until manually killed since the kernel thread
responsible has crashed. If such a program is run repeatedly, system resources
could eventually be exhausted, making this a vector for a denial of service
attack by unprivileged code.

This issue was previously raised in #11281, and a fix applied was to the
single ioctl that was reported as being problematic in that issue. To
further improve the situation, this change introduces a simple check in
zfs_check_input_nvpairs for the presence of either NV_UNIQUE_NAME or
NV_UNIQUE_NAME_TYPE flags and returns EINVAL if not present, which
prevents this issue from occuring for any of the non-legacy ioctls.

Signed-off-by: amy bones <amy@spookygirl.boo>